Plugins: fix streamer_tcp
